Fix Windows threads on MinGW
Because of the way CMake-shipped FindThreads.cmake works, under MinGW, both
CMAKE_USE_WIN32_THREADS_INIT and CMAKE_USE_PTHREADS_INIT are defined as true.
However, CMAKE_THREAD_LIBS_INIT is defined as "-pthread", and since libssh
determines the threading model by checking for pthreads first,
CMAKE_USE_WIN32_THREADS_INIT is ignored. One cannot simply uninstall
winpthreads, as some headers from there are needed for compilation. Who knew.

The solution is to simply check for Windows thread support first to use it.
Note that the preference for pthreads was removed as this ought to be a user's
decision, not a packager's. This commit DOES push Windows threads on the user,
but for my purposes this is fine.